+++ CHANGELOG.md
@@ -8,6 +8,7 @@ _Not yet on NuGet..._
 * Generate: Random seed uses `System.Random.Shared` on .NET platforms where it is available (#4217) @LeaFrock
 * Axes: Added `ClipLabel` option to prevent long labels from overlapping on very small plots (#4219) @drolevar
 * Plot: Improved performance when adding new plot objects by reducing complexity of color palette sampling (#4218) @0xfded @StendProg
+* Colormap: Added `GetColor()` overload to get color of an item in a collection with an option to sample from a range of the colormap
 
 ## ScottPlot 5.0.38
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2024-09-02_